requirements.txt
python-dotenv==1.0.0
requests==2.22.0
pandas==2.0.3